requirements.txt
altgraph==0.17.4
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
google-api-core==2.24.0
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httplib2==0.22.0
idna==3.8
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pandocfilters==1.5.1
paramiko==3.5.1
parse_pip_search==0.0.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.6.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
PyNaCl==1.5.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
pystray==0.19.5
python-dateutil==2.9.0.post0
pywin32==308
pywin32-ctypes==0.2.3
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rpds-py==0.23.1
rsa==4.9
selenium==4.29.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.4.0
tkhtmlview==0.3.1
tkinterdnd2==0.4.2
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.1
ttkbootstrap==1.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarg==0.1.9
